+++ redwood-treehouse-host/src/iosMain/kotlin/app/cash/redwood/treehouse/treehouseAppFactoryIos.kt
@@ -17,7 +17,6 @@ package app.cash.redwood.treehouse
 
 import app.cash.zipline.loader.ManifestVerifier
 import app.cash.zipline.loader.ZiplineHttpClient
-import kotlin.DeprecationLevel.HIDDEN
 import okio.FileSystem
 import okio.Path
 
@@ -27,31 +26,6 @@ public fun TreehouseAppFactory(
   manifestVerifier: ManifestVerifier,
   embeddedFileSystem: FileSystem? = null,
   embeddedDir: Path? = null,
-  cacheName: String = "zipline",
-  cacheMaxSizeInBytes: Long = 50L * 1024L * 1024L,
-  concurrentDownloads: Int = 8,
-  stateStore: StateStore = MemoryStateStore(),
-): TreehouseApp.Factory = TreehouseApp.Factory(
-  platform = IosTreehousePlatform(),
-  dispatchers = IosTreehouseDispatchers(),
-  httpClient = httpClient,
-  frameClockFactory = IosDisplayLinkClock,
-  manifestVerifier = manifestVerifier,
-  embeddedFileSystem = embeddedFileSystem,
-  embeddedDir = embeddedDir,
-  cacheName = cacheName,
-  cacheMaxSizeInBytes = cacheMaxSizeInBytes,
-  concurrentDownloads = concurrentDownloads,
-  stateStore = stateStore,
-)
-
-@Deprecated("", level = HIDDEN) // TODO Delete overload once 0.11.0 ships.
-@Suppress("FunctionName")
-public fun TreehouseAppFactory(
-  httpClient: ZiplineHttpClient,
-  manifestVerifier: ManifestVerifier,
-  embeddedDir: Path? = null,
-  embeddedFileSystem: FileSystem? = null,
   cacheName: String = "zipline",
   cacheMaxSizeInBytes: Long = 50L * 1024L * 1024L,
   concurrentDownloads: Int = 8,