+++ redwood-lazylayout-dom/src/commonMain/kotlin/app/cash/redwood/lazylayout/dom/HTMLLazyList.kt
@@ -22,53 +22,141 @@ import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.CrossAxisAlignment
 import app.cash.redwood.lazylayout.api.ScrollItemIndex
 import app.cash.redwood.lazylayout.widget.LazyList
+import app.cash.redwood.lazylayout.widget.LazyListScrollProcessor
+import app.cash.redwood.lazylayout.widget.LazyListUpdateProcessor
 import app.cash.redwood.lazylayout.widget.RefreshableLazyList
 import app.cash.redwood.ui.Margin
-import app.cash.redwood.widget.HTMLElementChildren
-import app.cash.redwood.widget.MutableListChildren
+import app.cash.redwood.widget.Widget
 import org.w3c.dom.Document
+import org.w3c.dom.Element
 import org.w3c.dom.HTMLDivElement
 import org.w3c.dom.HTMLElement
+import org.w3c.dom.get
 
 internal open class HTMLLazyList(document: Document) : LazyList<HTMLElement> {
   override var modifier: Modifier = Modifier
 
-  final override val placeholder = MutableListChildren<HTMLElement>()
-
   final override val value = document.createElement("div") as HTMLDivElement
+
+  private val processor = object : LazyListUpdateProcessor<HTMLElement, HTMLElement>() {
+    override fun insertRows(index: Int, count: Int) {
+    }
+
+    override fun deleteRows(index: Int, count: Int) {
+    }
+
+    override fun setContent(view: HTMLElement, content: Widget<HTMLElement>?) {
+      if (content != null) {
+        view.appendChild(content.value)
+      }
+    }
+  }
+
+  private val scrollProcessor = object : LazyListScrollProcessor() {
+    override fun contentSize(): Int = processor.size
+
+    override fun programmaticScroll(firstIndex: Int, animated: Boolean) {
+      TODO("Not yet implemented")
+    }
+  }
+
+  private val visibleSet = mutableSetOf<Element>()
+  private var highestEverVisibleIndex = 0
+
+  private val intersectionObserver = IntersectionObserver(
+    { entries, _ ->
+      entries.forEach { entry ->
+        if (entry.isIntersecting) {
+          visibleSet.add(entry.target)
+        } else {
+          visibleSet.remove(entry.target)
+        }
+      }
+
+      val highestVisibleIndex = items.widgets.indexOfLast { it.value in visibleSet }
+      if (highestVisibleIndex > highestEverVisibleIndex) {
+        highestEverVisibleIndex = highestVisibleIndex
+        // We currently won't unload any items that have been scrolled out of view
+        scrollProcessor.onUserScroll(firstIndex = 0, lastIndex = highestEverVisibleIndex)
+      }
+    },
+  )
 
   init {
     value.style.display = "flex"
   }
 
-  final override val items = HTMLElementChildren(value)
+  final override val items: Widget.Children<HTMLElement> = object : Widget.Children<HTMLElement> by processor.items {
+    override fun insert(index: Int, widget: Widget<HTMLElement>) {
+      processor.items.insert(index, widget)
+      intersectionObserver.observe(widget.value)
+
+      // Null element returned when index == childCount causes insertion at end.
+      val current = value.children[index]
+      value.insertBefore(widget.value, current)
+    }
+
+    override fun remove(index: Int, count: Int) {
+      for (i in index..<index + count) {
+        intersectionObserver.unobserve(value.children[i]!!)
+      }
+      processor.items.remove(index, count)
+
+      repeat(count) {
+        value.removeChild(value.children[index]!!)
+      }
+    }
+  }
+
+  final override val placeholder = processor.placeholder
 
   override fun width(width: Constraint) {
+    value.style.width = width.toCss()
   }
 
   override fun height(height: Constraint) {
+    value.style.height = height.toCss()
   }
 
   override fun margin(margin: Margin) {
+    value.style.apply {
+      marginInlineStart = margin.start.toPxString()
+      marginInlineEnd = margin.end.toPxString()
+      marginTop = margin.top.toPxString()
+      marginBottom = margin.bottom.toPxString()
+    }
   }
 
   override fun crossAxisAlignment(crossAxisAlignment: CrossAxisAlignment) {
   }
 
   override fun scrollItemIndex(scrollItemIndex: ScrollItemIndex) {
+    scrollProcessor.scrollItemIndex(scrollItemIndex)
   }
 
   override fun isVertical(isVertical: Boolean) {
-    value.style.flexDirection = if (isVertical) "column" else "row"
+    value.style.apply {
+      flexDirection = if (isVertical) "column" else "row"
+      if (isVertical) {
+        overflowY = "scroll"
+        removeProperty("overflowX")
+      } else {
+        overflowX = "scroll"
+        removeProperty("overflowY")
+      }
+    }
   }
 
   override fun onViewportChanged(onViewportChanged: (firstVisibleItemIndex: Int, lastVisibleItemIndex: Int) -> Unit) {
+    scrollProcessor.onViewportChanged(onViewportChanged)
   }
 
   override fun itemsBefore(itemsBefore: Int) {
+    processor.itemsBefore(itemsBefore)
   }
 
   override fun itemsAfter(itemsAfter: Int) {
+    processor.itemsAfter(itemsAfter)
   }
 }